fixed devices router logic

+++ src/routes/devises_router.ts
@@ -49,10 +49,6 @@ devicesRouter.delete('/', async (req, res)=>{
 
 devicesRouter.delete('/:id', async (req, res)=>{
     try {
-        if (!req.params.id) {
-            res.sendStatus(404);
-            return;
-        }
         const cookie: string = req.cookies.refreshToken
         const checkToken = await jwtService.verifyUserIdByRefreshToken(cookie)
         if (!checkToken) {
@@ -61,9 +57,14 @@ devicesRouter.delete('/:id', async (req, res)=>{
         }
         const checkId = await devicesService.findSessions(checkToken?.userId, new ObjectId(req.params.id))
         if (!checkId) {
+            res.sendStatus(404);
+            return;
+        }
+        if(checkId.deviceId.toString()!==req.params.id){
             res.sendStatus(403);
             return;
-        } else {
+        }
+        else {
             await devicesService.deleteOneSessionById(checkToken.deviceId)
             res.sendStatus(204);
             return;
